implement series and index nunique in IndexOpsMixin
